[Chore]#17 - 프로토콜 폴더 분리 재 커밋
