+++ ToDwoong/ToDwoong/Presentation/AddTodo/Protocols/AddTodoLocationPickerViewDelegate.swift
@@ -5,4 +5,6 @@
 //  Created by mirae on 3/5/24.
 //
 
-import Foundation
+protocol AddTodoLocationPickerViewDelegate: AnyObject {
+    func didTapConfirmAddress(_ address: String)
+}
