+++ ToDwoong/ToDwoong/Presentation/AddTodo/Protocols/DateTimePickerDelegate.swift
@@ -5,4 +5,6 @@
 //  Created by mirae on 3/5/24.
 //
 
-import Foundation
+protocol DateTimePickerDelegate: AnyObject {
+    func didPickDateOrTime(date: Date, mode: UIDatePicker.Mode)
+}
