+++ ToDwoong/ToDwoong/Presentation/AddTodo/Protocols/TitleCollectionViewCellDelegate.swift
@@ -5,4 +5,6 @@
 //  Created by mirae on 3/5/24.
 //
 
-import Foundation
+protocol TitleCollectionViewCellDelegate: AnyObject {
+    func titleCellDidEndEditing(_ text: String?)
+}
